Remove RecordIndexActionMenuBar and replace it with RecordIndexActionMenuButtons
